add FindMatchDecorationModel (#171857)
